e2e(#1397): one cap-patch verb, typed off the generated wire row
`adminPatchCaps` stood five times over five spec files. The bodies were
byte-identical; only the `caps` parameter diverged, three ways — and
measured, no caller depended on the divergence: unifying on the narrowest
of the three type-checks clean, with a bogus key still rejected.

The parameter keeps the two-term spelling one of the copies had, because
the two terms are two AXES rather than three values of one.
`Grappa.Admission` enforces the per-subject caps in a clause per
subject_kind, each reading the column its kind selects, over live
registry sessions; `max_per_ip` gets its own function, counts distinct
persisted subjects keyed on the session IP, and reads subject_kind as a
filter. The sentinel settles it: `nil` means unlimited on a per-subject
cap and means "use the operator default" on `max_per_ip`. The server's
own generated wire already groups them that way —
`cap_counts_changed` carries the two per-subject caps with their live
counts and no `max_per_ip`.

The type is DERIVED from `NetworksAdminWireT` rather than hand-spelled,
so the #767 drift gate keeps it aligned and a server-side rename fails
`tsc` here. It also makes the verb honest by construction: the endpoint
casts seven fields, and picking three is what stops a function named
`adminPatchCaps` from moving `visitor_enabled` — which the widest of the
three old copies would have allowed.

This adds the only import in `grappaApi.ts`, and it is type-only.
`wireTypes.ts` is generated and declares no imports itself, so it can
neither pull a fixture back in nor reach the solid-js dependency that
keeps `e2e/` from importing the rest of `src/`.
