+++ cicchetto/e2e/tests/u-2-admission-split.spec.ts
@@ -36,7 +36,12 @@
 // afterEach restores caps to permissive defaults so subsequent specs
 // see the seeder baseline.
 
-import { GRAPPA_BASE_URL, login, patchNetworkConnectionState } from "../fixtures/grappaApi";
+import {
+  adminPatchCaps,
+  GRAPPA_BASE_URL,
+  login,
+  patchNetworkConnectionState,
+} from "../fixtures/grappaApi";
 import {
   ADMIN_IDENTIFIER,
   ADMIN_PASSWORD,
@@ -46,26 +51,6 @@ import {
 import { expect, specUser, test } from "../fixtures/test";
 
 const SEED_CHANNEL = AUTOJOIN_CHANNELS[0];
-
-type CapDimension = "max_concurrent_user_sessions" | "max_concurrent_visitor_sessions";
-
-async function adminPatchCaps(
-  adminToken: string,
-  slug: string,
-  caps: Partial<Record<CapDimension | "max_per_ip", number | null>>,
-): Promise<void> {
-  const res = await fetch(`${GRAPPA_BASE_URL}/admin/networks/${encodeURIComponent(slug)}`, {
-    method: "PATCH",
-    headers: {
-      authorization: `Bearer ${adminToken}`,
-      "content-type": "application/json",
-    },
-    body: JSON.stringify(caps),
-  });
-  if (!res.ok) {
-    throw new Error(`adminPatchCaps: ${slug} → ${res.status} ${await res.text()}`);
-  }
-}
 
 // PATCH /networks/:slug {connection_state: "connected"} returning the
 // raw response so the matrix arm can assert status + body shape
