+++ cicchetto/e2e/tests/u-3-cap-honesty-mapping.spec.ts
@@ -42,7 +42,12 @@
 // afterEach restores caps to permissive defaults.
 
 import { expectShellReady, openAdminConsole } from "../fixtures/cicchettoPage";
-import { GRAPPA_BASE_URL, login, patchNetworkConnectionState } from "../fixtures/grappaApi";
+import {
+  adminPatchCaps,
+  GRAPPA_BASE_URL,
+  login,
+  patchNetworkConnectionState,
+} from "../fixtures/grappaApi";
 import {
   ADMIN_IDENTIFIER,
   ADMIN_PASSWORD,
@@ -54,26 +59,6 @@ import {
 import { expect, specUser, test } from "../fixtures/test";
 
 const SEED_CHANNEL = AUTOJOIN_CHANNELS[0];
-
-type CapKnob = "max_concurrent_user_sessions" | "max_concurrent_visitor_sessions" | "max_per_ip";
-
-async function adminPatchCaps(
-  adminToken: string,
-  slug: string,
-  caps: Partial<Record<CapKnob, number | null>>,
-): Promise<void> {
-  const res = await fetch(`${GRAPPA_BASE_URL}/admin/networks/${encodeURIComponent(slug)}`, {
-    method: "PATCH",
-    headers: {
-      authorization: `Bearer ${adminToken}`,
-      "content-type": "application/json",
-    },
-    body: JSON.stringify(caps),
-  });
-  if (!res.ok) {
-    throw new Error(`adminPatchCaps: ${slug} → ${res.status} ${await res.text()}`);
-  }
-}
 
 // /connect a network via the user-flow REST surface — returns the raw
 // Response so the arm can assert status + body shape without the
