+++ cicchetto/e2e/tests/ux-5-br-home-reconnect.spec.ts
@@ -37,6 +37,7 @@
 
 import { loginAs, waitForUserTopicReady } from "../fixtures/cicchettoPage";
 import {
+  adminPatchCaps,
   GRAPPA_BASE_URL,
   login,
   patchNetworkConnectionState,
@@ -51,26 +52,6 @@ import {
 import { expect, specNick, specUser, test } from "../fixtures/test";
 
 const SEED_CHANNEL = AUTOJOIN_CHANNELS[0];
-
-type CapKnob = "max_concurrent_user_sessions" | "max_concurrent_visitor_sessions" | "max_per_ip";
-
-async function adminPatchCaps(
-  adminToken: string,
-  slug: string,
-  caps: Partial<Record<CapKnob, number | null>>,
-): Promise<void> {
-  const res = await fetch(`${GRAPPA_BASE_URL}/admin/networks/${encodeURIComponent(slug)}`, {
-    method: "PATCH",
-    headers: {
-      authorization: `Bearer ${adminToken}`,
-      "content-type": "application/json",
-    },
-    body: JSON.stringify(caps),
-  });
-  if (!res.ok) {
-    throw new Error(`adminPatchCaps: ${slug} → ${res.status} ${await res.text()}`);
-  }
-}
 
 async function fetchChannels(token: string): Promise<Array<{ name: string; joined: boolean }>> {
   const res = await fetch(`${GRAPPA_BASE_URL}/networks/${NETWORK_SLUG}/channels`, {
